reactive-chat-service: refactoring

+++ reactive-chat-service/src/main/kotlin/com/ilchern/reactivechatservice/service/notifier/Notifier.kt
@@ -8,7 +8,7 @@ interface Notifier {
 
   val eventType: ChatEventType
 
-  fun notify(event: ChatEventEnvelope) : Mono<Int>
+  fun notify(event: ChatEventEnvelope): Mono<Long>
 
   fun extractMessageId(event: ChatEventEnvelope): String? {
     return event.payload
@@ -17,4 +17,3 @@ interface Notifier {
       ?.asText()
   }
 }
-
